Move files to endpoints/ dir
